eglib: simplify compiler checks
This removes a number of ./configure time compiler checks in favour of
compile time checks and also assumes you are building with a modern
compiler. For example we assume <stdint.h> and <inttypes.h> headers are
available.

These changes will make it easier to maintain the ability to build Cogl
on Android via an Android.mk file where it's not usual to use the build
system to check compiler features.