Add Android.mk based build support
This makes it possible to build Cogl for Android using ndk-build instead
of autotools. This is the typical way to build software for Android and
much simpler than using autotools given that there's no need to
dynamically probe the platform being built for. For ease of maintenance
this shares the list of cogl dependencies in cogl/Makefile.sources.