Tightly integrate ulib (was eglib)
Instead of maintaining optional support for using deps/eglib instead of
glib we now always use this code. This way the code is always being
tested so we are less likely to break support for building Cogl without
glib, and because the api no longer needs to be swappable we are free to
make Cogl specific changes to the api.

So that we can still support glib integration the eglib library has been
renamed to ulib and symbols are prefixed with 'u_' or 'U_' so that they
don't clash with glib.

GLib integration for GError and GType continues to be supported.