Remove support for CoglTextureRectangle
TextureRectangle used to be useful for supporting X11 texture from
pixmap on older ATI cards without full NPOT texture support and
theoretically it was there to help at times when it's preferable
to access a texture with non-normalized coordinates.

More recent versions of glsl now offer a way to sample texture with
non-normalized coordinates and so in practice this feature represents
quite a lot of code and special cases for non-normalized coordinates for
basically zero benefit.