kms-winsys: drop support for old gbm versions
Right now cogl tries to accomodate older gbm versions that
have the function gbm_bo_get_pitch() instead of the more
recently named gbm_bo_get_stride().

This adds an ugly #ifdef in the code. Furthermore, we are
going to rely on newer gbm for dma-buf support anyway.

This commit drops the #ifdef.

https://bugzilla.gnome.org/show_bug.cgi?id=753531